+++ test/modules/other-module/components/otherModule.ts
@@ -1,0 +1,3 @@
+import { SECRET_CREATE } from '@test/modules/secret-module'
+
+export const OTHER_MODULE = SECRET_CREATE