+++ test/modules/other-module/features/cool-feature/coolFeature.ts
@@ -1,0 +1,3 @@
+import { SECRET_CREATE } from '@test/modules/secret-module'
+
+const _test = SECRET_CREATE