+++ test/modules/secret-module/features/secret-feature/secretFeature.ts
@@ -1,0 +1,11 @@
+import { SHARED_CONST } from '@test/components/sharedFile'
+import { SECRET } from '@test/modules/secret-module/components/secret'
+import { SECRET_CREATE } from '@test/modules/secret-module/features/create/components/secretCreate'
+
+import { SECRET_FEATURE_2 } from './secretFeature2'
+
+export const SECRET_FEATURE = 'secret-create'
+const _test2 = SECRET_FEATURE_2
+const _test = SECRET
+const _test3 = SECRET_CREATE
+const _test4 = SHARED_CONST