+++ test/modules/secret-module/features/secret-feature/secretFeature2.ts
@@ -1,0 +1,1 @@
+export const SECRET_FEATURE_2 = 'secret-create'