+++ test/modules/secret-module/index.ts
@@ -1,0 +1,3 @@
+import { SECRET_CREATE } from './features/create/components/secretCreate'
+
+export { SECRET_CREATE }